third_party/blink/web_tests/external/wpt/html/webappapis/scripting/events/event-handler-all-global-events-expected.txt

This is a testharness.js-based test.
Found 8 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] onbeforematch: the content attribute must be compiled into a function as the corresponding property
  assert_equals: The onbeforematch property must be a function expected "function" but got "object"
[FAIL] onbeforematch: the content attribute must execute when an event is dispatched
  assert_true: Dispatching an event must run the code expected true got undefined
[FAIL] oncopy: must be on the appropriate locations for GlobalEventHandlers
  assert_true: Window has an own property named "oncopy" expected true got false
[FAIL] oncopy: the default value must be null
  assert_equals: The default value of the property is null for a Window instance expected (object) null but got (undefined) undefined
[FAIL] oncut: must be on the appropriate locations for GlobalEventHandlers
  assert_true: Window has an own property named "oncut" expected true got false
[FAIL] oncut: the default value must be null
  assert_equals: The default value of the property is null for a Window instance expected (object) null but got (undefined) undefined
[FAIL] onpaste: must be on the appropriate locations for GlobalEventHandlers
  assert_true: Window has an own property named "onpaste" expected true got false
[FAIL] onpaste: the default value must be null
  assert_equals: The default value of the property is null for a Window instance expected (object) null but got (undefined) undefined
Harness: the test ran to completion.
